Move RuleSetConfigProperty to formatting
