+++ detekt-formatting/src/main/kotlin/io/gitlab/arturbosch/detekt/formatting/RuleSetConfigProperty.kt
@@ -1,4 +1,4 @@
-package io.gitlab.arturbosch.detekt.api.internal
+package io.gitlab.arturbosch.detekt.formatting
 
 import io.gitlab.arturbosch.detekt.api.Config
 import kotlin.properties.ReadOnlyProperty
