+++ detekt-formatting/src/test/kotlin/io/gitlab/arturbosch/detekt/formatting/RuleSetConfigPropertySpec.kt
@@ -1,4 +1,4 @@
-package io.gitlab.arturbosch.detekt.api.internal
+package io.gitlab.arturbosch.detekt.formatting
 
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.RuleSet
